Added logs and health check endpoints

+++ Services/SubmissionService.cs
@@ -12,64 +12,75 @@ public class SubmissionService : ISubmissionService
     private readonly AppDbContext _context;
     private readonly ILogger<SubmissionService> _logger;
     private readonly IFileStorageService _fileStorageService;
-    private readonly RedisCacheSercvice _redisCacheSercvice;
+    private readonly RedisCacheService _redisCacheService;
     private readonly IMessagePublisher _messagePublisher;
 
  
-    public SubmissionService(AppDbContext context, ILogger<SubmissionService> logger, IFileStorageService fileStorageService, RedisCacheSercvice redisCacheSercvice, IMessagePublisher messagePublisher)
+    public SubmissionService(AppDbContext context, ILogger<SubmissionService> logger, IFileStorageService fileStorageService, RedisCacheService redisCacheService, IMessagePublisher messagePublisher)
     {
         _logger = logger;
         _context = context;
         _fileStorageService = fileStorageService;
-        _redisCacheSercvice = redisCacheSercvice;
+        _redisCacheService = redisCacheService;
         _messagePublisher = messagePublisher;
-        
     }
    
  
     public async Task<List<SubmissionResponse>> GetAll()
     {
+        _logger.LogDebug("Retrieving all submissions from database.");
         return await _context.Submissions.Select(t => new SubmissionResponse(t)).ToListAsync();
     }
  
     public async Task<SubmissionResponse> GetById(int Id)
     {
         string cacheKey = $"submission:{Id}";
-
-        SubmissionResponse? cachedSubmissionResponse = await _redisCacheSercvice.GetKeyAsync<SubmissionResponse>(cacheKey);
+        _logger.LogDebug("Checking Redis cache for key: {CacheKey}", cacheKey);
+        SubmissionResponse? cachedSubmissionResponse = await _redisCacheService.GetKeyAsync<SubmissionResponse>(cacheKey);
 
         if(cachedSubmissionResponse != default) {
-            _logger.LogInformation("Cache hit, Found the Submission with Id: {Id}", Id);
+            _logger.LogDebug("Cache hit. Found Submission with Id: {SubmissionId}", Id);
             return cachedSubmissionResponse;
         }
 
+        _logger.LogInformation("Cache miss. Fetching Submission from database for Id: {SubmissionId}", Id);
         Submission? submission = await _context.Submissions.FindAsync(Id);
         if(submission == null)
         {
-            _logger.LogInformation("Submission not found with {Id}", Id);
+            _logger.LogWarning("Record not found. Resource: {ResourceType}, Identifier: {Identifier}", "Submission", Id);
             throw new NotFoundException($"Submission not found with Id: {Id}");
         }
         SubmissionResponse submissionResponse = new SubmissionResponse(submission);
-        await _redisCacheSercvice.SetKeyAsync(cacheKey, submissionResponse);
+
+        _logger.LogDebug("Populating Redis cache key: {CacheKey} with Submission data.", cacheKey);
+        await _redisCacheService.SetKeyAsync(cacheKey, submissionResponse);
         return submissionResponse;
     }
  
     public async Task<SubmissionResponse> Create(CreateSubmissionRequest createSubmissionRequest)
     {
         bool taskAssignmentExists = await _context.TaskAssignments.AnyAsync(t => t.Id == createSubmissionRequest.TaskAssignmentId);
-        if(!taskAssignmentExists) throw new NotFoundException($"TaskAssignment not found with Id: {createSubmissionRequest.TaskAssignmentId}");
+        if(!taskAssignmentExists)
+        {
+            _logger.LogWarning("Record not found. Resource: {ResourceType}, Identifier: {Identifier}", "TaskAssignment", createSubmissionRequest.TaskAssignmentId);
+            throw new NotFoundException($"TaskAssignment not found with Id: {createSubmissionRequest.TaskAssignmentId}");
+        }
  
         Submission submission = new Submission(createSubmissionRequest);
         await _context.Submissions.AddAsync(submission);
         await _context.SaveChangesAsync();
-        _logger.LogInformation("Successfully Created Submission: {Id} ", submission.Id);
+        _logger.LogInformation("Submission event: {ActionEvent} occurred for SubmissionId: {SubmissionId}", "Created", submission.Id);
         return new SubmissionResponse(submission);
     }
  
     public async Task<SubmissionFileResponse> UploadFile(int userId, int submissionId, CreateSubmissionFileRequest createSubmissionFileRequest)
     {
         bool submission = await _context.Submissions.AnyAsync(s => s.Id == submissionId);
-        if(!submission) throw new NotFoundException($"TaskAssignment not found with Id: {submissionId}");
+        if(!submission)
+        {
+            _logger.LogWarning("Record not found. Resource: {ResourceType}, Identifier: {Identifier}", "Submission", submissionId);
+            throw new NotFoundException($"TaskAssignment not found with Id: {submissionId}");
+        }
         
         var correlationId = Guid.NewGuid();
  
@@ -79,7 +90,7 @@ public class SubmissionService : ISubmissionService
         Stream stream = file.OpenReadStream();
         string? userName = await _context.Users.Where(u => u.Id == userId).Select(u => u.UserName).FirstOrDefaultAsync();
         await _fileStorageService.SaveAsync(generatedFileName, stream);
-        string CheckSum = _fileStorageService.GetChecksum(stream);
+        string CheckSum = _fileStorageService.GetChecksum(await _fileStorageService.OpenReadAsync(generatedFileName));
         SubmissionFile submissionFile = new()
         {
             SubmissionId = submissionId,
@@ -94,22 +105,33 @@ public class SubmissionService : ISubmissionService
         await _context.SubmissionFiles.AddAsync(submissionFile);
         await _context.SaveChangesAsync();
 
-        // try
-        // {
+        _logger.LogInformation("Submission file metadata stored. FileId: {FileId}, StoragePathName: {GeneratedName}", submissionFile.Id, generatedFileName);
+
+        try
+        {
             var message = new SubmissionProcessingRequested(
                 MessageId: Guid.NewGuid(),
                 CorrelationId: correlationId,
-                SubmissionId: Guid.NewGuid(),
+                SubmissionId: submissionId,
                 FileId: submissionFile.Id,
                 RequestedAt: DateTime.UtcNow.ToUtcSecondPrecision()
             );
 
-            await _messagePublisher.PublishAsync(message);
             
-        // } catch(Exception)
-        // {
-        //     throw new ServiceUnavailableException("Database update completed, but background processing pipeline is temporarily offline.");
-        // }
+            await _messagePublisher.PublishAsync(message);
+            ProcessingJob processingJob = new ProcessingJob
+            {
+                Status = "Queued",
+                CorrelationId = correlationId
+            };
+            await _context.ProcessingJobs.AddAsync(processingJob);
+            await _context.SaveChangesAsync();
+            _logger.LogInformation("Successfully queued background job validation task. CorrelationId: {CorrelationId}", correlationId);
+        } catch(Exception ex)
+        {
+            _logger.LogError(ex, "Metadata stored successfully, but broker submission dispatch failed. CorrelationId: {CorrelationId}", correlationId);
+            throw new ServiceUnavailableException("Database update completed, but background processing pipeline is temporarily offline.");
+        }
  
         return new SubmissionFileResponse
         {
@@ -127,16 +149,32 @@ public class SubmissionService : ISubmissionService
 
     public async Task<FileStream> DownloadFile(int submissionFileId)
     {
-        SubmissionFile? submissionFile = await _context.SubmissionFiles.FindAsync(submissionFileId) ?? throw new NotFoundException($"SubmissionFile not found with Id: {submissionFileId}");
+        SubmissionFile? submissionFile = await _context.SubmissionFiles.FindAsync(submissionFileId);
+        if (submissionFile == null)
+        {
+            _logger.LogWarning("Record not found. Resource: {ResourceType}, Identifier: {Identifier}", "SubmissionFile", submissionFileId);
+            throw new NotFoundException($"SubmissionFile not found with Id: {submissionFileId}");
+        }
+
+        _logger.LogInformation("Processing download authorization for FileId: {FileId}", submissionFileId);
         return await _fileStorageService.OpenReadAsync(submissionFile.GeneratedStorageName);
     }
 
     public async Task<bool> DeleteFile(int submissionFileId)
     {
-        SubmissionFile? submissionFile = await _context.SubmissionFiles.FindAsync(submissionFileId) ?? throw new NotFoundException($"SubmissionFile not found with Id: {submissionFileId}");
+        SubmissionFile? submissionFile = await _context.SubmissionFiles.FindAsync(submissionFileId);
+        if (submissionFile == null)
+        {
+            _logger.LogWarning("Record not found. Resource: {ResourceType}, Identifier: {Identifier}", "SubmissionFile", submissionFileId);
+            throw new NotFoundException($"SubmissionFile not found with Id: {submissionFileId}");
+        }
+        _logger.LogInformation("Initiating structural file deleting routine. FileId: {FileId}, StoragePath: {GeneratedName}", 
+            submissionFileId, submissionFile.GeneratedStorageName);
         await _fileStorageService.DeleteAsync(submissionFile.GeneratedStorageName);
         _context.Remove(submissionFile);
         await _context.SaveChangesAsync();
+
+        _logger.LogInformation("Submission file asset deleted from database and local storage. FileId: {FileId}", submissionFileId);
         return true;
     }
 }
